refactor: UserMessageの配置を変更
