+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/common/UserMessage.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.common
 
 import androidx.annotation.StringRes
